% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distances.R
\name{find_all_nodes}
\alias{find_all_nodes}
\title{Find all nodes within search_radius_m of long and lat of multiple coordinates}
\usage{
find_all_nodes(
  coord_df,
  node_data,
  longitude_col = "Longitude",
  latitude_col = "Latitude",
  search_radius_m = 2000
)
}
\arguments{
\item{coord_df}{Data frame with longitude and latitude of points from which to find NAPTAN nodes}

\item{node_data}{NAPTAN node data from}

\item{longitude_col}{Longitude column name}

\item{latitude_col}{Latitude column name}

\item{search_radius_m}{Search radius in meters}
}
\value{
Returns data frame of NAPTAN nodes within search_radius_m of all given coordinates
}
\description{
Find all nodes within search_radius_m of long and lat of multiple coordinates
}
\examples{
node_data <- get_single_region_nodes("West Midlands")

coord_df <- data.frame(
  "LONG" = c(-1.9060413, -1.7802142),
  "LAT" = c(52.4803994, 52.4115384)
)

nodes_in_300m <- find_all_nodes(
  coord_df, 
  node_data,
  longitude_col = "LONG",
  latitude_col = "LAT",
  search_radius_m = 300 # metres
)

nodes_in_300m
}
