% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naptan_api.R
\name{get_multi_region_nodes}
\alias{get_multi_region_nodes}
\title{Call NAPTAN API to get all nodes for one or more regions.}
\usage{
get_multi_region_nodes(region_array)
}
\arguments{
\item{region_array}{Array of NAPTAN region. Use \code{naptanr::lookup_atco_codes()} to find allowed region names.}
}
\value{
Returns data frame of NAPTAN nodes within given region
}
\description{
Call NAPTAN API to get all nodes for one or more regions.
}
\examples{

region_array <- c("West Midlands", "Worcestershire", "Staffordshire", "Warwickshire")

get_single_region_nodes(region_array)
}
