% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distances.R
\name{get_nearest_node}
\alias{get_nearest_node}
\title{Get single NAPTAN node closest to any of the given coordinates}
\usage{
get_nearest_node(
  coord_df,
  region_array,
  longitude_col = "longitude",
  latitude_col = "latitude",
  search_radius_m = 2000
)
}
\arguments{
\item{coord_df}{Data frame with longitude and latitude of points from which to find NAPTAN nodes}

\item{region_array}{Array of NAPTAN regions. Use \code{naptanr::lookup_atco_codes()} to find allowed region names.}

\item{longitude_col}{Longitude column name}

\item{latitude_col}{Latitude column name}

\item{search_radius_m}{Search radius in meters}
}
\value{
Returns a single row data frame with the NAPTAN node closest to all given coordinates
}
\description{
Get single NAPTAN node closest to any of the given coordinates
}
\examples{

coord_df <- data.frame(
"LONG" = c(-1.9060413, -1.7802142),
"LAT" = c(52.4803994, 52.4115384)
)

nearest_node <- get_nearest_node(
  coord_df,
  c("West Midlands"),
  longitude_col = "LONG",
  latitude_col = "LAT",
  search_radius_m = 200
)

nearest_node
}
