% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distances.R
\name{get_nodes}
\alias{get_nodes}
\title{Find all nodes within search_radius_m of long and lat (Single point)}
\usage{
get_nodes(long, lat, node_data, search_radius_m = 2000)
}
\arguments{
\item{long}{Longitude}

\item{lat}{Latitude}

\item{node_data}{NAPTAN node data from get_single_region_nodes() or get_multi_region_nodes()}

\item{search_radius_m}{Search radius in meters}
}
\value{
Returns data frame of NAPTAN nodes within search_radius_m of given coordinates
}
\description{
Find all nodes within search_radius_m of long and lat (Single point)
}
\examples{
node_data <- get_single_region_nodes("West Midlands")

nodes_in_300m <- get_nodes(
   -1.9060413,
   52.4803994,
   node_data,
   search_radius_m = 300 # metres
)

nodes_in_300m
}
