% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naptan_api.R
\name{get_single_region_nodes}
\alias{get_single_region_nodes}
\title{Call NAPTAN API to get all nodes for a given region.}
\usage{
get_single_region_nodes(region_name)
}
\arguments{
\item{region_name}{NAPTAN region. Use \code{naptanr::lookup_atco_codes()} to find allowed region names.}
}
\value{
Returns data frame of NAPTAN nodes within given region
}
\description{
Call NAPTAN API to get all nodes for a given region.
}
\examples{

get_single_region_nodes("West Midlands")

}
