% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ukos_to_lonlat}
\alias{ukos_to_lonlat}
\title{Convert UKOS eastings and northings to longitude and latitude}
\usage{
ukos_to_lonlat(df, easting_col = "Easting", northing_col = "Northing")
}
\arguments{
\item{df}{Data frame containing UKOS eastings and northings}

\item{easting_col}{Eastings column name}

\item{northing_col}{Northings column name}
}
\value{
Returns original data frame with additional longitude and latitude columns
}
\description{
Convert UKOS eastings and northings to longitude and latitude
}
\examples{

df <- data.frame(
"EAST" = c(436403, 436400),
"NORTH" = c(280955, 280967)
)

ukos_to_lonlat(
  df,
  easting_col = "EAST",
  northing_col = "NORTH"
)
}
